ESP32: Use untildaytime endpoint so sleep until 6
Use the previously introduced endpoint on the flask side of things
to sleep until morning as soon as we hit 22:00. That way we don't
waste a bunch of power at night pointlessly refreshing an empty
graph.

This adds two more config.h values, API_URL (which should point
towards the untildaytime endpoit) and TIMEZONE (which should be
a standard timezone name, e.g. "Europe/Zurich").