flask: Use the max_over_time for day_energy
Sometimes the inverter wakes up at the end of the day and decides
it's a new day when posting its data, which awkwardly results in a
total energy production of reportedly 0.

Ideally we'd sum those into some grand total, but it's easier to
just take the maximum.

Solar eclipses might break this.